Update SwerveModule.java
Looking to use CTRE SRX Max as a absolut encoder, which will attache to the RoboRio, and this post said to use a DutyCycleEncoder, which sounds like it will count the ticks, which will tell where things are at in the 360 degree pointing of the wheel.  https://www.chiefdelphi.com/t/how-do-i-use-ctre-mag-encoder-as-an-absolute-encoder/397394/3

Actualy I think woudl be easier to just get a CTRE that can comunicate over CAN a so called CAN ENcoder. https://store.ctr-electronics.com/cancoder/